[BugFix] Attach configured MCP servers in AgenticNode base so every subagent honors its mcp selection
