+++ datus/agent/node/gen_sql_agentic_node.py
@@ -106,18 +106,10 @@ class GenSQLAgenticNode(AgenticNode):
             input_data=input_data,
             agent_config=agent_config,
             tools=tools or [],
-            mcp_servers={},  # Initialize empty, will setup after parent init
+            mcp_servers={},
             scope=scope,
             is_subagent=is_subagent,
             session_id=session_id,
-        )
-
-        # Initialize MCP servers based on configuration (after node_config is available)
-        self.mcp_servers = self._setup_mcp_servers()
-
-        # Debug: Log final MCP servers assignment
-        logger.debug(
-            f"GenSQLAgenticNode final mcp_servers: {len(self.mcp_servers)} servers - {list(self.mcp_servers.keys())}"
         )
 
         # Setup tools based on configuration (includes subagent task tool wiring)
@@ -516,61 +508,6 @@ class GenSQLAgenticNode(AgenticNode):
                 logger.warning(f"Method '{method_name}' not found in {tool_type}")
         except Exception as e:
             logger.error(f"Failed to setup {tool_type}.{method_name}: {e}")
-
-    def _setup_mcp_server_from_config(self, server_name: str) -> Optional[Any]:
-        """Setup MCP server from {agent.home}/conf/.mcp.json using mcp_manager."""
-        try:
-            from datus.tools.mcp_tools.mcp_manager import MCPManager
-
-            # Use MCPManager to get server config
-            mcp_manager = MCPManager(agent_config=self.agent_config)
-            server_config = mcp_manager.get_server_config(server_name)
-
-            if not server_config:
-                logger.warning(f"MCP server '{server_name}' not found in configuration")
-                return None
-
-            # Create server instance using the manager
-            server_instance, details = mcp_manager._create_server_instance(server_config)
-
-            if server_instance:
-                logger.debug(f"Added MCP server '{server_name}' from configuration: {details}")
-                return server_instance
-            else:
-                error_msg = details.get("error", "Unknown error")
-                logger.warning(f"Failed to create MCP server '{server_name}': {error_msg}")
-                return None
-
-        except Exception as e:
-            logger.error(f"Failed to setup MCP server '{server_name}' from config: {e}")
-            return None
-
-    def _setup_mcp_servers(self) -> Dict[str, Any]:
-        """Set up MCP servers based on configuration."""
-        mcp_servers = {}
-
-        config_value = self.node_config.get("mcp", "")
-        if not config_value:
-            return mcp_servers
-
-        mcp_server_names = [p.strip() for p in config_value.split(",") if p.strip()]
-
-        for server_name in mcp_server_names:
-            try:
-                server = self._setup_mcp_server_from_config(server_name)
-                if server:
-                    mcp_servers[server_name] = server
-
-            except Exception as e:
-                logger.error(f"Failed to setup MCP server '{server_name}': {e}")
-
-        logger.debug(f"Setup {len(mcp_servers)} MCP servers: {list(mcp_servers.keys())}")
-
-        # Debug: Log detailed info about each server
-        for name, server in mcp_servers.items():
-            logger.debug(f"MCP server '{name}': type={type(server)}, instance={server}")
-
-        return mcp_servers
 
     def _get_available_tool_names(self) -> list[str]:
         """Return native tool names plus discoverable MCP tool names."""
